Try again? I think I got a better algo for this
